refactor: introduce AbstractMultipleDependentResourceIT
