+++ operator-framework/src/test/java/io/javaoperatorsdk/operator/MultipleDependentResourceIT.java
@@ -1,61 +1,12 @@
 package io.javaoperatorsdk.operator;
 
-import java.time.Duration;
-import java.util.stream.IntStream;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.RegisterExtension;
-
-import io.fabric8.kubernetes.api.model.ConfigMap;
-import io.fabric8.kubernetes.api.model.ObjectMetaBuilder;
-import io.javaoperatorsdk.operator.junit.LocalOperatorExtension;
-import io.javaoperatorsdk.operator.sample.multipledependentresource.MultipleDependentResourceConfigMap;
-import io.javaoperatorsdk.operator.sample.multipledependentresource.MultipleDependentResourceCustomResource;
 import io.javaoperatorsdk.operator.sample.multipledependentresource.MultipleDependentResourceReconciler;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.awaitility.Awaitility.await;
+class MultipleDependentResourceIT
+    extends AbstractMultipleDependentResourceIT<MultipleDependentResourceReconciler> {
 
-class MultipleDependentResourceIT {
-
-  public static final String TEST_RESOURCE_NAME = "multipledependentresource-testresource";
-  @RegisterExtension
-  LocalOperatorExtension operator =
-      LocalOperatorExtension.builder().withReconciler(MultipleDependentResourceReconciler.class)
-          .waitForNamespaceDeletion(true)
-          .build();
-
-  @Test
-  void twoConfigMapsHaveBeenCreated() {
-    MultipleDependentResourceCustomResource customResource = createTestCustomResource();
-    operator.create(MultipleDependentResourceCustomResource.class, customResource);
-
-    var reconciler = operator.getReconcilerOfType(MultipleDependentResourceReconciler.class);
-
-    await().pollDelay(Duration.ofMillis(300))
-        .until(() -> reconciler.getNumberOfExecutions() <= 1);
-
-    IntStream.of(MultipleDependentResourceReconciler.FIRST_CONFIG_MAP_ID,
-        MultipleDependentResourceReconciler.SECOND_CONFIG_MAP_ID).forEach(configMapId -> {
-          ConfigMap configMap =
-              operator.get(ConfigMap.class, customResource.getConfigMapName(configMapId));
-          assertThat(configMap).isNotNull();
-          assertThat(configMap.getMetadata().getName())
-              .isEqualTo(customResource.getConfigMapName(configMapId));
-          assertThat(configMap.getData().get(MultipleDependentResourceConfigMap.DATA_KEY))
-              .isEqualTo(String.valueOf(configMapId));
-        });
+  @Override
+  protected Class<MultipleDependentResourceReconciler> getReconcilerClass() {
+    return MultipleDependentResourceReconciler.class;
   }
-
-  public MultipleDependentResourceCustomResource createTestCustomResource() {
-    MultipleDependentResourceCustomResource resource =
-        new MultipleDependentResourceCustomResource();
-    resource.setMetadata(
-        new ObjectMetaBuilder()
-            .withName(TEST_RESOURCE_NAME)
-            .withNamespace(operator.getNamespace())
-            .build());
-    return resource;
-  }
-
 }
